+++ lib/bautom.rb
@@ -1,6 +1,0 @@
-require "bautom/version"
-
-module Bautom
-  class Error < StandardError; end
-  # Your code goes here...
-end