+++ spec/booking_automation_spec.rb
@@ -1,0 +1,5 @@
+RSpec.describe BookingAutomation do
+  it 'has a version number' do
+    expect(BookingAutomation::VERSION).not_to be nil
+  end
+end