+++ spec/spec_helper.rb
@@ -1,9 +1,12 @@
-require "bundler/setup"
-require "bautom"
+require 'bundler/setup'
+require 'webmock/rspec'
+
+require 'booking_automation'
+Dir[File.join(Dir.pwd, 'spec/support/**/*.rb')].each { |f| require f }
 
 RSpec.configure do |config|
   # Enable flags like --only-failures and --next-failure
-  config.example_status_persistence_file_path = ".rspec_status"
+  config.example_status_persistence_file_path = '.rspec_status'
 
   # Disable RSpec exposing methods globally on `Module` and `main`
   config.disable_monkey_patching!