+++ CHANGELOG.md
@@ -7,6 +7,14 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
-## [0.1.0] - 2019-06-27
+## [0.1.0] - 2019-06-30
 
 ### Added
+
+- Added ability to list properties
+- Added ability to fetch individual property
+- Added ability to show bookings for a property
+
+### Changed
+
+### Removed